+++ docs/CHANGELOG.md
@@ -8,9 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.33] - 2026-08-04
 
+### Added
+
+- The stop conditions of `notifications listen` can be set in the configuration file
+
 ### Changed
 
 - Option `--advertise-url` of `notifications listen` renamed to `--register-url-full`
+- The configuration keys that only `notifications listen` uses moved under their own
+  subsection
 
 ### Fixed